+++ src/view/mod.rs
@@ -1,4 +1,3 @@
-pub mod debug;
 pub mod layout;
 pub mod listen;
 pub mod session;